requirements.txt
fastapi
uvicorn[standard]
kafka-python
httpx
python-dotenv
